move TimeStampDivider to src/components/DateStampDivider. convert to typescript
